--- src/main/java/com/b3/ddarelro/domain/comment/controller/CommentListReq.java
@@ -1,5 +1,0 @@
-package com.b3.ddarelro.domain.comment.controller;
-
-public record CommentListReq(Long cardId) {
-
-}
