+++ src/main/java/com/b3/ddarelro/domain/comment/dto/CommentListReq.java
@@ -1,0 +1,5 @@
+package com.b3.ddarelro.domain.comment.dto;
+
+public record CommentListReq(Long cardId) {
+
+}
