+++ src/main/java/com/b3/ddarelro/domain/comment/dto/CommentListRes.java
@@ -1,4 +1,4 @@
-package com.b3.ddarelro.domain.comment.controller;
+package com.b3.ddarelro.domain.comment.dto;
 
 import lombok.Builder;
